refactor: separate AddItem
