+++ src/routes/Settings/URLsManager/Item/index.ts
@@ -1,9 +1,5 @@
 // Copyright (C) 2017-2024 Smart code 203358507
 
 import Item from './Item';
-import AddItem from './AddItem';
 
-export {
-    Item,
-    AddItem,
-};
+export default Item;
